add task and testcase logics

+++ informatics-repository/src/main/java/ge/freeuni/informatics/repository/task/ITaskRepository.java
@@ -6,6 +6,6 @@ public interface ITaskRepository {
 
     void addTask(Task task);
 
-    Task getTask(Long taskId);
+    Task getTask(int taskId);
     
 }
